udisks2: don't mount non-removable and non-thumb devices
We don't want devices which are not marked as removable or as a thumb
drive to be ever selected as candidate for the auto-mounter. The udisks2
API documentation states that some USB storages devices are marked with
MediaRemovable=true even if they don't have any removable media. We need
to make sure that we correctly check both, the MediaRemovable and the
Removable property in our detection logic. In addition to that we then
only want thumb drives to be allowed for the auto-mounter.